Added New Function
Added a function called _quickScanOpen which does a quick scan of all the ports of a target until it finds an open one. If it finds an open one, it returns the port number; if no open port found, it returns False. Added a call to this function in the fingerOS function.